+++ packages/crdt_lf/test/handler/text/handler_test.dart
@@ -284,8 +284,7 @@ void main() {
         // check convergence
         expect(text1.value.length, 3);
         expect(text1.value, equals(text3.value));
-        expect(text1.value, equals(text3.value));
-
+        
         expect(text1.value, contains('A'));
         expect(text1.value, contains('B'));
         expect(text1.value, contains('C'));
@@ -310,12 +309,6 @@ void main() {
         expect(text1.value, equals(text3.value));
         expect(text2.value, equals(text3.value));
         final convergedValue = text1.value;
-        expect(
-          convergedValue == 'YBAX' ||
-              convergedValue == 'BCAX' ||
-              convergedValue == 'CBAX',
-          isTrue,
-        );
 
         // take snapshot and sync
         final snapshot1 = doc1.takeSnapshot();